% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{buildModelFuncs}
\alias{buildModelFuncs}
\title{Build c++ particle filtering code for your R session, and save it as a 
pfrmodel object.}
\usage{
buildModelFuncs(myDir, model_name, save = TRUE)
}
\arguments{
\item{myDir}{directory with your three code files (i.e. model header, model source and export code)}

\item{model_name}{your model name. Must be in all lowercase, and be a substring of the above-mentioned filenames}

\item{save}{if TRUE, then save the compiled object (file name suffix is " _pfrmodel.rds")}
}
\value{
a pfrmodel object
}
\description{
Build c++ particle filtering code for your R session, and save it as a 
pfrmodel object.
}
\examples{
\dontrun{
# compile everything from scratch
svol_lev <- buildModelFuncs("~/Desktop", "svol_leverage", save = TRUE)

# ...or load a previously-compiled pfrmodel object
svol_lev <- readRDS("~/Desktop/svol_leverage_pfrmodel.rds")

# then use your model's log-likelihood and filtering functions
svol_lev_funcs <- unPackFunctions(svol_lev)
svol_lev_funcs$svol_leverage_bswc_approx_LL(rnorm(100), c(.9, 0.0, 1.0, -.2))
svol_lev_funcs$svol_leverage_bswc_approx_filt(rnorm(100), c(.9, 0.0, 1.0, -.2))
}
}
