% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_creators.R
\name{createAuxiliaryFilterTemplate}
\alias{createAuxiliaryFilterTemplate}
\title{Create file templates for an auxiliary particle filter model.}
\usage{
createAuxiliaryFilterTemplate(modname, saveFile = TRUE, fileDir = NULL)
}
\arguments{
\item{modname}{name of model in all lowercase}

\item{saveFile}{TRUE if you want to save the file, FALSE otherwise}

\item{fileDir}{where to save files. If NULL, saves to ./src/}
}
\value{
NULL if saving a file, otherwise a list with three character vectors
}
\description{
If you want to run an auxiliary particle filter, this function creates 
the three required C++ files associated with your state-space model.
}
\examples{
# files pop up in rstudio
createAuxiliaryFilterTemplate("coolmod") 

# does not pop up in rstudio
myCon <- file("myFile.cpp", "wb")
threeCodes <- createAuxiliaryFilterTemplate("mod", saveFile = F)
expCodeStr <- threeCodes$export_code
writeLines(expCodeStr , myCon )
close(myCon)

}
