\name{pfr-package}
\alias{pfr-package}
\alias{pfr}
\docType{package}
\title{
\packageTitle{pfr}
}
\description{
\packageDescription{pfr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pfr}
\packageIndices{pfr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{pfr}

Maintainer: \packageMaintainer{pfr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
