% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{unPackFunctions}
\alias{unPackFunctions}
\title{Unpack functions from a pfrmodel object.}
\usage{
unPackFunctions(pfModel)
}
\arguments{
\item{pfModel}{your pfrmodel object}
}
\value{
an Rcpp module with your functions from your export code
}
\description{
Unpack functions from a pfrmodel object.
}
\examples{
\dontrun{
# compile everything from scratch
svol_lev <- buildModelFuncs("~/Desktop", "svol_leverage", save = TRUE)

# ...or load a previously-compiled pfrmodel object
svol_lev <- readRDS("~/Desktop/svol_leverage_pfrmodel.rds")

# then use your model's log-likelihood and filtering functions
svol_lev_funcs <- unPackFunctions(svol_lev)
svol_lev_funcs$svol_leverage_bswc_approx_LL(rnorm(100), c(.9, 0.0, 1.0, -.2))
svol_lev_funcs$svol_leverage_bswc_approx_filt(rnorm(100), c(.9, 0.0, 1.0, -.2))
}
}
